// packages/source-deploy-retrieve/src/client/types.ts
export type RecordId = string;

export enum ComponentStatus {
  Created = 'Created',
  Changed = 'Changed',
  Unchanged = 'Unchanged',
  Deleted = 'Deleted',
  Failed = 'Failed',
}
// ...
interface FileResponseBase {
  fullName: string;
  type: string;
  filePath?: string;
}

interface FileResponseSuccess extends FileResponseBase {
  state: Exclude<ComponentStatus, ComponentStatus.Failed>;
}

interface FileResponseFailure extends FileResponseBase {
  state: ComponentStatus.Failed;
  lineNumber?: number;
  columnNumber?: number;
  error: string;
  problemType: 'Warning' | 'Error';
}

export type FileResponse = FileResponseSuccess | FileResponseFailure;

export interface MetadataTransferResult {
  response: MetadataRequestStatus;
  components: ComponentSet;
  getFileResponses(): FileResponse[];
}

export interface SourceApiResult {
  success: boolean;
}

export interface AsyncResult {
  id: RecordId;
}

export interface SourceDeployResult extends SourceApiResult {
  id: RecordId;
  components?: ComponentDeployment[];
  status: RequestStatus | ToolingDeployStatus;
}

export enum RequestStatus {
  Pending = 'Pending',
  InProgress = 'InProgress',
  Succeeded = 'Succeeded',
  SucceededPartial = 'SucceededPartial',
  Failed = 'Failed',
  Canceling = 'Canceling',
  Canceled = 'Canceled',
}

export interface MetadataRequestStatus {
  id: string;
  status: RequestStatus;
  success: boolean;
  done: boolean;
}
// ...
export type RetrieveRequest = {
  apiVersion: string;
  packageNames?: string[];
  singlePackage?: boolean;
  specificFiles?: string[];
  unpackaged: {
    types: PackageTypeMembers[];
  };
};

export type RetrieveMessage = { fileName: string; problem: string };

enum ManageableState {
  Beta = 'beta',
  Deleted = 'deleted',
  Deprecated = 'deprecated',
  DeprecatedEditable = 'deprecatedEditable',
  Installed = 'installed',
  InstalledEditable = 'installedEditable',
  Released = 'released',
  Unmanaged = 'unmanaged',
}

export type FileProperties = {
  createdById: string;
  createdByName: string;
  createdDate: string;
  fileName: string;
  fullName: string;
  id: string;
  lastModifiedById: string;
  lastModifiedByName: string;
  lastModifiedDate: string;
  manageableState?: ManageableState;
  namespacePrefix?: string;
  type: string;
};

/**
 * Raw response returned from a checkRetrieveStatus call to the Metadata API
 */
export interface MetadataApiRetrieveStatus {
  done: boolean;
  fileProperties: FileProperties | FileProperties[];
  id: string;
  status: RequestStatus;
  success: boolean;
  messages?: RetrieveMessage[] | RetrieveMessage;
  /** `base64` encoded string */
  zipFile: string;
}

// ------------------------------------------------
// Client options
// ------------------------------------------------

export interface PackageOption {
  /**
   * The name of the package to retrieve.
   */
  name: string;
  /**
   * The directory where the retrieved package source should be
   * converted. If this is not specified the directory will
   * default to `<process.cwd()>/PackageOption.name`.
   */
  outputDir?: SourcePath;
}

export type PackageOptions = string[] | PackageOption[];

export interface RetrieveExtractOptions {
  /**
   * Top most directory within the retrieved zip file.
   * E.g., `unpackaged` for unpackaged source, or the name of the
   * package for retrieved package source.
   */
  zipTreeLocation: string;
  /**
   * The directory where the retrieved source should be converted.
   * This is `RetrieveOptions.output` for unpackaged source, and
   * `PackageOption.outputDir` for packaged source.
   */
  outputDir: SourcePath;
}

export interface RetrieveOptions {
  /**
   * The directory to retrieve components to. If `merge: true`, components are only
   * retrieved to `output` if there wasn't a component to merge with.
   */
  output: SourcePath;
  /**
   * Whether or not to merge and replace input components with the retrieved versions.
   */
  merge?: boolean;
  /**
   * A list of package names to retrieve, or package names and their retrieval locations.
   */
  packageOptions?: PackageOptions;
}